+++ build.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env sh
+
+set -e
+
+"`dirname "$0"`/docker/build.sh"