+++ lint.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env sh
+
+set -e
+
+js-lint "`dirname "$0"`"